% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikBetaProportion}
\alias{llikBetaProportion}
\title{log likelihood of the BetaProportion distribution and its derivatives (from stan)}
\usage{
llikBetaProportion(x, mu, kappa, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{mu}{distribution parameter}

\item{kappa}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the BetaProportion distribution and its derivatives (from stan)
}
