% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llik.R
\name{llikCauchy}
\alias{llikCauchy}
\title{log likelihood of Cauchy distribution and it's derivatives (from stan)}
\usage{
llikCauchy(x, location = 0, scale = 1, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{location, scale}{location and scale parameters.}

\item{full}{Add the data frame showing x, mean, sd as well as the
fx and derivatives}
}
\value{
data frame with \code{fx} for the log pdf value of with
\code{dLocation} and \code{dScale} that has the derivatives with respect to the parameters at
the observation time-point
}
\description{
log likelihood of Cauchy distribution and it's derivatives (from stan)
}
\examples{

x <- seq(-3, 3, length.out = 21)

llikCauchy(x, 0, 1)

llikCauchy(x, 3, 1, full=TRUE)

}
\author{
Matthew L. Fidler
}
