% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llik.R
\name{llikExp}
\alias{llikExp}
\title{log likelihood and derivatives for exponential distribution}
\usage{
llikExp(x, rate, full = FALSE)
}
\arguments{
\item{x}{variable that is distributed by exponential distribution}

\item{rate}{vector of rates.}

\item{full}{Add the data frame showing x, mean, sd as well as the
fx and derivatives}
}
\value{
data frame with \code{fx} for the log pdf value of with \code{dRate}
that has the derivatives with respect to the \code{rate} parameter
the observation time-point
}
\description{
log likelihood and derivatives for exponential distribution
}
\examples{

llikExp(1, 1:3)

llikExp(1, 1:3, full=TRUE)

}
\author{
Matthew L. Fidler
}
