% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikFrechet}
\alias{llikFrechet}
\title{log likelihood of the Frechet distribution and its derivatives (from stan)}
\usage{
llikFrechet(x, alpha, sigma, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{alpha}{distribution parameter}

\item{sigma}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the Frechet distribution and its derivatives (from stan)
}
