% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llik.R
\name{llikGeom}
\alias{llikGeom}
\title{log likelihood and derivatives for Geom distribution}
\usage{
llikGeom(x, prob, full = FALSE)
}
\arguments{
\item{x}{variable distributed by a geom distribution}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}

\item{full}{Add the data frame showing x, mean, sd as well as the
fx and derivatives}
}
\value{
data frame with \code{fx} for the log pdf value of with \code{dProb}
that has the derivatives with respect to the \code{prob} parameters at
the observation time-point
}
\description{
log likelihood and derivatives for Geom distribution
}
\examples{

llikGeom(1:10, 0.2)

}
\author{
Matthew L. Fidler
}
