% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikGumbel}
\alias{llikGumbel}
\title{log likelihood of the Gumbel distribution and its derivatives (from stan)}
\usage{
llikGumbel(x, mu, beta, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{mu}{distribution parameter}

\item{beta}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the Gumbel distribution and its derivatives (from stan)
}
