% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikInvChisq}
\alias{llikInvChisq}
\title{log likelihood of the InvChisq distribution and its derivatives (from stan)}
\usage{
llikInvChisq(x, nu, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{nu}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the InvChisq distribution and its derivatives (from stan)
}
