% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikInvGamma}
\alias{llikInvGamma}
\title{log likelihood of the InvGamma distribution and its derivatives (from stan)}
\usage{
llikInvGamma(x, alpha, beta, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{alpha}{distribution parameter}

\item{beta}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the InvGamma distribution and its derivatives (from stan)
}
