% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikLnorm}
\alias{llikLnorm}
\title{log likelihood of the Lnorm distribution and its derivatives (from stan)}
\usage{
llikLnorm(x, meanlog, sdlog, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{meanlog}{distribution parameter}

\item{sdlog}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the Lnorm distribution and its derivatives (from stan)
}
