% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikLogis}
\alias{llikLogis}
\title{log likelihood of the Logis distribution and its derivatives (from stan)}
\usage{
llikLogis(x, location, scale, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{location}{distribution parameter}

\item{scale}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the Logis distribution and its derivatives (from stan)
}
