% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llik.R
\name{llikNbinom}
\alias{llikNbinom}
\title{Calculate the log likelihood of the negative binomial function (and its derivatives)}
\usage{
llikNbinom(x, size, prob, full = FALSE)
}
\arguments{
\item{x}{Number of successes}

\item{size}{Dispersion parameter; may be non-integer}

\item{prob}{probability of success}

\item{full}{Add the data frame showing x, mean, sd as well as the
fx and derivatives}
}
\value{
data frame with \code{fx} for the pdf value of with
\code{dProb} that has the derivatives with respect to the parameters at
the observation time-point
}
\description{
Calculate the log likelihood of the negative binomial function (and its derivatives)
}
\examples{

llikNbinom(46:54, 100, 0.5)

llikNbinom(46:54, 100, 0.5, TRUE)

}
\author{
Matthew L. Fidler
}
