% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikPareto}
\alias{llikPareto}
\title{log likelihood of the Pareto distribution and its derivatives (from stan)}
\usage{
llikPareto(x, yMin, alpha, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{yMin}{distribution parameter}

\item{alpha}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the Pareto distribution and its derivatives (from stan)
}
