% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikParetoType2}
\alias{llikParetoType2}
\title{log likelihood of the ParetoType2 distribution and its derivatives (from stan)}
\usage{
llikParetoType2(x, mu, lambda, alpha, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{mu}{distribution parameter}

\item{lambda}{distribution parameter}

\item{alpha}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the ParetoType2 distribution and its derivatives (from stan)
}
