% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llik.R
\name{llikPois}
\alias{llikPois}
\title{log-likelihood for the Poisson distribution}
\usage{
llikPois(x, lambda, full = FALSE)
}
\arguments{
\item{x}{non negative integers}

\item{lambda}{non-negative means}

\item{full}{Add the data frame showing x, mean, sd as well as the
fx and derivatives}
}
\value{
data frame with \code{fx} for the pdf value of with
\code{dLambda} that has the derivatives with respect to the parameters at
the observation time-point
}
\description{
log-likelihood for the Poisson distribution
}
\author{
Matthew L. Fidler
}
