% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikRayleigh}
\alias{llikRayleigh}
\title{log likelihood of the Rayleigh distribution and its derivatives (from stan)}
\usage{
llikRayleigh(x, sigma, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{sigma}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the Rayleigh distribution and its derivatives (from stan)
}
