% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llikNew.R
\name{llikScaledInvChisq}
\alias{llikScaledInvChisq}
\title{log likelihood of the ScaledInvChisq distribution and its derivatives (from stan)}
\usage{
llikScaledInvChisq(x, nu, sigma, full = FALSE)
}
\arguments{
\item{x}{Observation}

\item{nu}{distribution parameter}

\item{sigma}{distribution parameter}

\item{full}{return the input parameters alongside the result}
}
\value{
data frame with \code{fx} and the derivative of each parameter
}
\description{
log likelihood of the ScaledInvChisq distribution and its derivatives (from stan)
}
