% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llik.R
\name{llikUnif}
\alias{llikUnif}
\title{log likelihood and derivatives for Unif distribution}
\usage{
llikUnif(x, alpha, beta, full = FALSE)
}
\arguments{
\item{x}{variable distributed by a uniform distribution}

\item{alpha}{is the lower limit of the uniform distribution}

\item{beta}{is the upper limit of the distribution}

\item{full}{Add the data frame showing x, mean, sd as well as the
fx and derivatives}
}
\value{
data frame with \code{fx} for the log pdf value of with \code{dProb}
that has the derivatives with respect to the \code{prob} parameters at
the observation time-point
}
\description{
log likelihood and derivatives for Unif distribution
}
\examples{

llikUnif(1, -2, 2)

}
\author{
Matthew L. Fidler
}
